% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read data from file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{input string}
}
\value{
data frame tbl with data read from provided file
    filename.  File filename is expected to be in .csv format.
    Function is stopped if the provided
    filename does not exist.
}
\description{
Reads data from a csv file and returns a data frame tbl object.
}
\examples{

#fars_read(my_file)

}
