% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Convert a vector of years to a list of filenames and imports them}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{vector of integer or string}
}
\value{
list of string
}
\description{
Converts a vector years to a list of filenames using the make_filename
  function.  Produces errors for invalid years in the input vector.
}
\examples{

#fars_read_years(c(1993,1995,2001,2016))

#fars_read_years(c("1993",1995,"2001",2016))

}
