% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Organizes data by year and month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{vector of years, integer or character}
}
\value{
data frame
}
\description{
Reads data with year and month components, returns a count
  of data entries for each month in each year, and organizes
  the return data frame using a separate column for
  each year.
}
\examples{

fars_summarize_years(c("1993",1995,"2001",2016))

}
