% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make a filename from a given year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{integer or string}
}
\value{
string, specialized filename containing year
}
\description{
Create a filename from a given year which includes the
  year in the correct locationin the name.
}
\examples{

make_filename(1992)

make_filename("1992")

}
